Update tiles/map using "timestamps"
When a map is updated, its updateTime is incremented, causing all owners
to also update since their mapUpdateTime field will be out of sync with
the map.

Once this is done, the owner then checks tiles in a bounding box and
sees if their modTime differs from the owner's TileView. If this
differs, then any objects unknown to the owner are send to the client,
and any known objects are updated.