Implement managing and sending tile lighting
This isn't properly mixed in with map local lighting or any sort of
global "outdoor" lighting state.

This honestly requires some thought... perhaps if the map is marked as
outdoor, and a given tile has a free shot to the "sky" it is considered
outdoors, whilst otherwise it'll use 100% local tile lights?