Add generic embeddable "feature" for inventory
This goes against our current patterns, but it might be wise to use a
composable structure based on embedding structs to provide
certain features to object. This simplifies object organization and
there doesn't seem to be a downside, besides having to also
convert other features to work in a similar manner.